[C] Upgrade wavesurfer
