+++ client/package.json
@@ -52,6 +52,7 @@
     "@emotion/server": "^11.11.0",
     "@emotion/styled": "^11.14.0",
     "@loadable/component": "^5.16.4",
+    "@wavesurfer/react": "^1.0.9",
     "actioncable": "~5.0.7",
     "autolinker": "^4.1.0",
     "buffer": "^6.0.3",
@@ -122,7 +123,7 @@
     "tus-js-client": "^1.8.0",
     "use-resize-observer": "^8.0.0",
     "uuid": "https://registry.npmjs.org/uuid/-/uuid-2.0.2.tgz",
-    "wavesurfer": "^1.3.4"
+    "wavesurfer.js": "^7.9.1"
   },
   "devDependencies": {
     "@babel/core": "^7.26.9",
